Reduce global scope, use cached currentCity
CurrentCity is cached within ForecastPanel to avoid "flickering" when the component reloads.
The component is reloaded once a minute, or when its props change. However, by default,
the component initializes with an empty forecast data.

In order to avoid that, the component's data is initialized as the "last received" value,
until loadCity replaces it.

Signed-off-by: Balint Erdosi <erdosib@gmail.com>