Clear interval on ForecastPanelCreated
On ForecastPanelCreated an interval is created, that checks the weather every minute.
As the funciton is bound to the component, and the interval is registered globally,
everytime the component is created, a new interval was created as well.

This commit saves the interval to the cache, and attempts to clear the preexisting one,
when the component is created.

Signed-off-by: Balint Erdosi <erdosib@gmail.com>